application/requirements/base.txt
django==1.11.4
django-environ==0.4.3
django-anymail==0.11.1
django-compressor==2.1.1
django-allauth==0.32.0
django-filter==1.0.4
djangorestframework==3.6.3
djangorestframework-jwt==1.11.0
celery==4.1.0
PyMySQL==0.7.11
SQLAlchemy==1.1.10
